Custom app now installable
The app will now install as a custom app using the generated link in
partners. The home page uses app bridge to construct a link with a
signed query string to the proxied portal. Tests are still incomplete
for the oauth process, but the basics are there. Next steps are to
configure and create the webhooks.