Added shopify registry
Now webhooks can be registered. In this commit the app/uninstalled
webhook is registered and works to clean up shopify_sessions and remove
stored registered webhooks.

Some more tests added for the webhooks.

A script directory added, this is for development to test things out. It
includes scripts to query endpoint apis.