Invalid value used as WeakMap key
I don't think this a fault with crank. However to make the app display
at all I needed to prevent minification with the vite build. The error
is still displayed in the inspector but at least it is no longer a show
topper. By removing redundant listeners from date and variant selector
then it reduced the number of errors by two.